agent-template/requirements.txt
# Core runtime
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
pyyaml>=6.0.1

# LLM providers (install the one(s) you use)
openai>=1.30.0
anthropic>=0.26.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
